Extract class into own file.
